add user mapping and edit functionality

+++ src/main/java/com/example/url_shortner/service/UrlService.java
@@ -1,10 +1,15 @@
 package com.example.url_shortner.service;
 
 
+import com.example.url_shortner.dto.OriginalUrlDto;
 import com.example.url_shortner.dto.ShortCodeResponseDto;
+import com.example.url_shortner.exceptions.NoSuchUserFoundException;
 import com.example.url_shortner.exceptions.ResourceNotFoundException;
+import com.example.url_shortner.exceptions.DuplicateShortCodeException;
 import com.example.url_shortner.model.UrlModel;
+import com.example.url_shortner.model.User;
 import com.example.url_shortner.repository.UrlRepository;
+import com.example.url_shortner.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
@@ -13,6 +18,7 @@ import org.springframework.stereotype.Service;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
+import java.time.LocalDateTime;
 import java.util.Base64;
 import java.util.Optional;
 
@@ -23,32 +29,64 @@ public class UrlService {
     private static final int SHORT_CODE_LENGTH = 7;
     private final UrlRepository urlRepository;
 
-    public UrlService(UrlRepository urlRepository){
+    private final UserRepository userRepository;
+
+    public UrlService(UrlRepository urlRepository,UserRepository userRepository){
         this.urlRepository = urlRepository;
+        this.userRepository = userRepository;
     }
 
-    public ResponseEntity<?> createShortUrl(String url){
+    @Transactional
+    public ResponseEntity<?> createShortUrl(OriginalUrlDto urlDto, String apiKey){
 
-//        Optional<UrlModel> existingUrl = urlRepository.findByOriginalUrl(url);
-//
-//        if(existingUrl.isPresent()){
-//            return new ResponseEntity<ShortCodeResponseDto>(new ShortCodeResponseDto(existingUrl.get().getShortCode()),
-//                    HttpStatus.CONFLICT);
-//        }
+        Optional<User> user = userRepository.findByApiKey(apiKey);
+
+        if(user.isEmpty()){
+            throw new NoSuchUserFoundException("No user with this api key is present",HttpStatus.NOT_FOUND);
+        }
 
         try{
 
-            String short_code = generateUniqueShortCode(url);
+            if(urlDto.getShortCode() != null && !urlDto.getShortCode().trim().isEmpty()){
+                String short_code = urlDto.getShortCode().trim();
+                Optional<UrlModel> existingUrlModel = urlRepository.findByShortCode(short_code);
 
-            UrlModel urlModel = new UrlModel();
+                if(existingUrlModel.isPresent()){
+                    throw new DuplicateShortCodeException(" The short code provided by you already exists",HttpStatus.BAD_REQUEST);
+                }else{
+                    String url = urlDto.getUrl();
 
-            urlModel.setOriginalUrl(url);
-            urlModel.setShortCode(short_code);
-            urlModel.setNumVisited(0L);
 
-            urlRepository.save(urlModel);
+                    UrlModel urlModel = new UrlModel();
 
-            return new ResponseEntity<ShortCodeResponseDto>(new ShortCodeResponseDto(short_code), HttpStatus.CREATED);
+                    urlModel.setOriginalUrl(url);
+                    urlModel.setShortCode(short_code);
+                    urlModel.setNumVisited(0);
+                    urlModel.setUser(user.get());
+                    urlModel.setExpiryDate(urlDto.getExpiryDate());
+
+                    urlRepository.save(urlModel);
+
+                    return new ResponseEntity<ShortCodeResponseDto>(new ShortCodeResponseDto(short_code), HttpStatus.CREATED);
+                }
+            }else{
+                String url = urlDto.getUrl();
+                String short_code = generateUniqueShortCode(url);
+
+                UrlModel urlModel = new UrlModel();
+
+                urlModel.setOriginalUrl(url);
+                urlModel.setShortCode(short_code);
+                urlModel.setNumVisited(0);
+                urlModel.setUser(user.get());
+                urlModel.setExpiryDate(urlDto.getExpiryDate());
+
+                urlRepository.save(urlModel);
+
+                return new ResponseEntity<ShortCodeResponseDto>(new ShortCodeResponseDto(short_code), HttpStatus.CREATED);
+            }
+
+
         }catch(DataAccessException e){
             return new ResponseEntity<String>(e.getMessage(),HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -67,11 +105,45 @@ public class UrlService {
     }
 
     @Transactional
-    public void deleteUrl(String shortCode){
+    public String makeUrlInactive(String shortCode, String apiKey, LocalDateTime expiryDate){
+        Optional<User> user = userRepository.findByApiKey(apiKey);
+
+        if(user.isEmpty()){
+            throw new NoSuchUserFoundException(" No user with this api key is found " ,HttpStatus.BAD_REQUEST);
+        }
+
+        if(apiKey!= null && !apiKey.equals(user.get().getApiKey())){
+            throw new RuntimeException(" Not authorized to change the url " + HttpStatus.UNAUTHORIZED);
+        }
+
+        Optional<UrlModel> existingUrlModel = urlRepository.findByShortCode(shortCode);
+
+        if(existingUrlModel.isEmpty()){
+            throw new ResourceNotFoundException(" No url with short code " + shortCode + " is present" , HttpStatus.BAD_REQUEST);
+        }
+
+        existingUrlModel.get().setExpiryDate(expiryDate);
+
+        urlRepository.save(existingUrlModel.get());
+
+        return "success";
+    }
+
+    @Transactional
+    public void deleteUrl(String shortCode,String api_key){
         Optional<UrlModel> url = urlRepository.findByShortCode(shortCode);
 
         if(url.isEmpty()){
             throw new ResourceNotFoundException(" No code with " + shortCode + " value exists " , HttpStatus.NOT_FOUND);
+        }
+        Optional<User> user = userRepository.findById(url.get().getUser().getId());
+
+        if(user.isEmpty()){
+            throw new NoSuchUserFoundException(" No user found " , HttpStatus.NOT_FOUND);
+        }
+
+        if(!user.get().getApiKey().equals(api_key)){
+            throw new NoSuchUserFoundException(" The api key is not associated with the owner of code " , HttpStatus.BAD_REQUEST);
         }
 
         urlRepository.deleteByShortCode(shortCode);
